make tanstack the new fireclient
