+++ apps/fireclient/.eslintrc.cjs
