--- apps/fireclient/Dockerfile
@@ -1,5 +1,0 @@
-FROM oven/bun AS base
-WORKDIR /app
-COPY . .
-RUN bun install --frozen-lockfile --production
-CMD ["bun", "run", "dev"]
