+++ apps/fireclient/shell.nix
