--- apps/fireclient/src/app/auth/page.tsx
@@ -1,9 +1,0 @@
-import { AuthForm } from "@/components/auth-form"
-
-export default function AuthPage() {
-  return (
-    <main>
-      <AuthForm />
-    </main>
-  )
-}
