--- apps/fireclient/src/app/server/[address]/[channel]/page.tsx
@@ -1,7 +1,0 @@
-export default function ChannelPage() {
-  return (
-    <main>
-      <h1>Channel Page</h1>
-    </main>
-  )
-}
