--- apps/fireclient/src/app/server/[address]/page.tsx
@@ -1,8 +1,0 @@
-export default function ServerHomePage({ params }: { params: { address: string } }) {
-  return (
-    <main>
-      <h1>Server Home Page</h1>
-      <p>Address: {params.address}</p>
-    </main>
-  )
-}
