+++ apps/fireclient/src/components/devconsole.tsx
