+++ apps/fireclient/src/components/sidebar.tsx
@@ -1,14 +1,58 @@
-import { RiMenuFold3Line } from "react-icons/ri";
-import Link from "next/link";
+import { RiMenuFold3Line, RiMenuUnfold3Line } from "react-icons/ri";
 import { FaHashtag, FaHouse, FaLinkSlash } from "react-icons/fa6"
 import { IdentitySwitcher } from "@/components/identity-switcher"
+import { connectionStore } from "@/lib/store"
+import { useStore } from "@tanstack/react-store";
+import { handleMaybe } from "maybe";
+import { Link } from "@tanstack/react-router"
 
-export function Sidebar() {
+
+export function SidebarLayout({ children }: Readonly<{ children: React.ReactNode }>) {
+  const connection = useStore(connectionStore)
+  let address = ""
+  handleMaybe(connection, (conn) => {
+    address = conn.address
+  }, () => {})
+
+  return (
+    <div className="drawer lg:drawer-open">
+      <input id="my-drawer-2" type="checkbox" className="drawer-toggle" />
+      <main className="drawer-content h-screen">
+        <Topbar address={address} />
+
+        {children}
+      </main>
+      <div className="drawer-side">
+        <label htmlFor="my-drawer-2" aria-label="close sidebar" className="drawer-overlay"></label>
+        <div className="menu bg-base-200 text-base-content min-h-full w-80 p-4">
+          <Sidebar address={address} />
+        </div>
+      </div>
+    </div>
+  )
+}
+
+
+
+// only shows up on small screens
+export function Topbar({ address }: { address: string}) {
+  return (
+    <div className="w-full flex flex-row bg-base-300 lg:hidden">
+      <label htmlFor="my-drawer-2" className="m-2 btn btn-circle btn-outline drawer-button">
+        <RiMenuUnfold3Line size="1.5em" />
+      </label>
+      <IdentitySwitcher address={address} className="ml-auto max-w-60" />
+    </div>
+  )
+}
+
+
+export function Sidebar({ address }: { address: string }) {
   // TODO: show the channel list if you're connected
   return (
     <>
       <div className="flex flex-row">
-        <IdentitySwitcher className="hidden lg:block" />
+        <IdentitySwitcher address={address} className="hidden lg:block" />
         <label htmlFor="my-drawer-2" className="ml-auto m-2 btn btn-circle btn-outline drawer-button lg:hidden">
           <RiMenuFold3Line size="1.5em" />
         </label>
@@ -60,10 +104,9 @@ type QuickLinkProps = {
 }
 
 function QuickLink(props: QuickLinkProps) {
-  "use client"
   return (
     <li className="w-full">
-      <Link href={props.to} className="flex flex-row items-center">
+      <Link to={props.to} className="flex flex-row items-center">
         {props.icon}
         <p className="font-bold">{props.text}</p>
       </Link>
@@ -98,7 +141,7 @@ type ChannelProps = {
 function Channel(props: ChannelProps) {
   return (
     <li>
-      <Link className="flex flex-row" href={props.id}>
+      <Link className="flex flex-row" to={`./${props.id}`}>
         <FaHashtag />
         <p>{props.name}</p>
       </Link>
