--- apps/fireclient/src/lib/address.ts
@@ -1,13 +1,0 @@
-export function urlFromAddress(address: string): string {
-  if (address.startsWith("localhost")) {
-    return `http://${address}`
-  }
-  return `https://${address}`
-}
-
-export function socketFromAddress(address: string): string {
-  if (address.startsWith("localhost")) {
-    return `ws://${address}/socket`
-  }
-  return `wss://${address}/socket`
-}
