+++ apps/fireclient/src/routes/__root.tsx
