+++ apps/fireclient/src/routes/auth.tsx
