+++ apps/fireclient/src/routes/server.tsx
