+++ apps/fireclient/src/routes/server/$address/index.tsx
