+++ apps/fireclient/src/vite-env.d.ts
