--- apps/fireclient/tailwind.config.ts
@@ -1,16 +1,0 @@
-import type { Config } from "tailwindcss";
-
-const config: Config = {
-  content: [
-    "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
-    "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
-    "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
-  ],
-  plugins: [
-    require("daisyui")
-  ],
-  daisyui: {
-    themes: ["dark"]
-  }
-};
-export default config;
