+++ apps/fireclient/tsconfig.app.json
