+++ apps/fireclient/vite.config.ts
