--- apps/tanstack-demo/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "tanstack-demo",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@tanstack/react-store": "^0.5.5",
-    "@tanstack/store": "^0.5.5",
-    "@types/node": "^20.14.11",
-    "daisyui": "^4.12.10",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-icons": "^5.2.1"
-  },
-  "devDependencies": {
-    "@tanstack/router-devtools": "^1.45.4",
-    "@tanstack/router-plugin": "^1.45.3",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.13.1",
-    "@typescript-eslint/parser": "^7.13.1",
-    "@vitejs/plugin-react-swc": "^3.5.0",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^8.57.0",
-    "eslint-plugin-react-hooks": "^4.6.2",
-    "eslint-plugin-react-refresh": "^0.4.7",
-    "postcss": "^8.4.39",
-    "tailwindcss": "^3.4.6",
-    "typescript": "^5.2.2",
-    "vite": "^5.3.1"
-  }
-}
